Rspec

+++ db/migrate/20250625052710_remove_devise_fields_from_users.rb
@@ -1,7 +1,7 @@
 class RemoveDeviseFieldsFromUsers < ActiveRecord::Migration[7.2]
   def change
-    remove_column :users, :reset_password_token, :string
-    remove_column :users, :reset_password_sent_at, :datetime
-    remove_column :users, :remember_created_at, :datetime
+    remove_column :users, :reset_password_token, :string if column_exists?(:users, :reset_password_token)
+    remove_column :users, :reset_password_sent_at, :datetime if column_exists?(:users, :reset_password_sent_at)
+    remove_column :users, :remember_created_at, :datetime if column_exists?(:users, :remember_created_at)
   end
 end
